Semana 14 Exercicios - Testes

+++ src/semana14/exercicios/exercicio1/Aquario.java
@@ -1,33 +1,33 @@
 package semana14.exercicios.exercicio1;
 
 public class Aquario {
-    float altura, largura, comprimento;
+    double altura, largura, comprimento;
 
-    public Aquario(float altura, float largura, float comprimento){
+    public Aquario(double altura, double largura, double comprimento){
         setAltura(altura);
         setLargura(largura);
         setComprimento(comprimento);
     }
 
-    public Aquario(float valorUnico){
+    public Aquario(double valorUnico){
         setAltura(valorUnico);
         setLargura(valorUnico);
         setComprimento(valorUnico);
     }
 
-    public float getAltura(){
+    public double getAltura(){
         return altura;
     }
 
-    public float getLargura(){
+    public double getLargura(){
         return largura;
     }
 
-    public float getComprimento(){
+    public double getComprimento(){
         return comprimento;
     }
 
-    public void setAltura(float altura){
+    public void setAltura(double altura){
         if (altura < 0.0) {
             throw new IllegalArgumentException("Altura inválido: Altura não pode ser negativo");
          } else if (altura == 0.0){
@@ -36,7 +36,7 @@ public class Aquario {
         this.altura = altura;
     }
 
-    public void setLargura(float largura){
+    public void setLargura(double largura){
         if (largura < 0.0) {
             throw new IllegalArgumentException("Largura inválido: Largura não pode ser negativo");
          } else if (largura == 0.0){
@@ -45,7 +45,7 @@ public class Aquario {
         this.largura = largura;
     }
 
-    public void setComprimento(float comprimento){
+    public void setComprimento(double comprimento){
         if (comprimento < 0.0) {
             throw new IllegalArgumentException("Comprimento inválido: Comprimento não pode ser negativo");
          } else if (comprimento == 0.0){
@@ -54,11 +54,11 @@ public class Aquario {
         this.comprimento = comprimento;
     }
 
-    public float calcularVolume() {
+    public double calcularVolume() {
         return (altura*largura*comprimento)/1000;
     }
     
-    public float calcularPotenciaTermostato(float temperaturaAmbiente, float temperaturaDesejada) {
+    public double calcularPotenciaTermostato(double temperaturaAmbiente, double temperaturaDesejada) {
         return calcularVolume() * 0.05f * (temperaturaAmbiente - temperaturaDesejada);
     }
     
@@ -66,11 +66,11 @@ public class Aquario {
         return "Filtragem mínima: " + calcularFiltragemMinima() + "\n Filtragem Máxima: " + calcularFiltragemMaxima();
     }
 
-    public float calcularFiltragemMinima() {
+    public double calcularFiltragemMinima() {
         return calcularVolume() * 2;
     }
     
-    public float calcularFiltragemMaxima() {
+    public double calcularFiltragemMaxima() {
         return calcularVolume() * 3;
     }
 }
